Fix docker run not previewing environment variables.
This acted somewhat flakey and seemed to only work when the environment
variables were edited. This is wrapped up in my recent change to always
use text inputs to edit the values, but even in the previous incarnation
it didn't always work. I'm confused how this works when the DL isn't
visible due to tabs, but the other tabs all seem to preview properly
with the :visible selector.